requirements.txt
streamlit==1.30.0
pillow==10.0.0
numpy==1.26.0
opencv-python-headless==4.8.1.78
